Fix: Missing color references in layouts
With Change-Id: I09db1208f7f7a6637aa89fdafdb40bd4eeb6a5bf there was a
color defined in colors.xml but the reference to it in layouts was
not included. Additionally I added a background for favorites view
that was missing in single panel view for tablets, as well as two
more colors to fix the call detail background being tied to the
blank background triggerd by the proximity sensor, and to allow for
more flexibility with theming.

Change-Id: I9ccab7cc0e8fc8b54c6b7cdcc6665dc19410c996